Change the utest to use the new kernel.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/scalfmm/scalfmm/trunk@301 2616d619-271b-44dc-8df4-d4a8f33a7222

+++ Src/Utils/FTic.hpp
@@ -81,7 +81,7 @@ public:
 #else
         timeval t;
         gettimeofday(&t, NULL);
-        return (t.tv_sec) + (t.tv_usec/1000000.0);
+        return double(t.tv_sec) + (double(t.tv_usec)/1000000.0);
 #endif
     }
 };
